Move memory allocation/deallocation into rust

+++ storage.h
@@ -44,8 +44,8 @@ extern const char *str_ref(const char *);
 extern Var memory_usage(void);
 
 extern void myfree(void *where, Memory_Type type);
-extern void *mymalloc(unsigned size, Memory_Type type);
-extern void *myrealloc(void *where, unsigned size, Memory_Type type);
+extern void *mymalloc(size_t size, Memory_Type type);
+extern void *myrealloc(void *where, size_t size, Memory_Type type);
 
 static inline void    /* XXX was extern, fix for non-gcc compilers */
 free_str(const char *s)
